+++ changelog.md
@@ -1,3 +1,9 @@
 # Changelog
 
-Base data inserted successfully.
+**./src/export-sql-files.js**
+* Added SQL export tasks for the remaining tables:
+	* SupportWorkerOtherLanguages
+	* SupportWorkerChecks
+	* SupportWorkerPersonality
+	* *[...]*
+	* SupportWorkerOther